requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amazon-ion==0.13.0
    # via kestra
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
babel==2.17.0
    # via mkdocs-material
backrefs==5.8
    # via mkdocs-material
bcrypt==4.3.0
    # via passlib
bidict==0.23.1
    # via python-socketio
black==25.1.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
brotli==1.1.0
    # via geventhttpclient
certifi==2025.4.26
    # via
    #   geventhttpclient
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   flask
    #   mkdocs
    #   uvicorn
colorama==0.4.6
    # via mkdocs-material
configargparse==1.7.1
    # via
    #   locust
    #   locust-cloud
coverage[toml]==7.8.2
    # via pytest-cov
cryptography==45.0.3
    # via python-jose
docker==7.1.0
    # via -r requirements.in
ecdsa==0.19.1
    # via python-jose
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flake8==7.2.0
    # via -r requirements.in
flask==3.1.1
    # via
    #   flask-cors
    #   flask-login
    #   locust
flask-cors==6.0.0
    # via locust
flask-login==0.6.3
    # via locust
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
gevent==24.11.1
    # via
    #   geventhttpclient
    #   locust
    #   locust-cloud
geventhttpclient==2.3.3
    # via locust
ghp-import==2.1.0
    # via mkdocs
greenlet==3.2.2
    # via
    #   gevent
    #   playwright
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
hf-xet==1.1.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
huggingface-hub==0.32.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   mkdocs
    #   mkdocs-material
    #   torch
joblib==1.5.1
    # via scikit-learn
kestra==0.22.0
    # via -r requirements.in
kuzu==0.10.0
    # via -r requirements.in
locust==2.37.5
    # via -r requirements.in
locust-cloud==1.21.8
    # via locust
markdown==3.8
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mkdocs
    #   werkzeug
mccabe==0.7.0
    # via flake8
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mkdocs==1.6.1
    # via
    #   -r requirements.in
    #   mkdocs-material
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-material==9.6.14
    # via -r requirements.in
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via locust
mypy-extensions==1.1.0
    # via black
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   scikit-learn
    #   scipy
    #   transformers
packaging==25.0
    # via
    #   black
    #   huggingface-hub
    #   mkdocs
    #   pytest
    #   transformers
paginate==0.5.7
    # via mkdocs-material
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   black
    #   mkdocs
pillow==11.2.1
    # via sentence-transformers
platformdirs==4.3.8
    # via
    #   black
    #   locust-cloud
    #   mkdocs-get-deps
playwright==1.52.0
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
prometheus-client==0.22.0
    # via -r requirements.in
psutil==7.0.0
    # via locust
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycodestyle==2.13.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyee==13.0.0
    # via playwright
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via mkdocs-material
pymdown-extensions==10.15
    # via mkdocs-material
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==6.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   ghp-import
    #   kestra
python-engineio==4.12.1
    # via python-socketio
python-jose[cryptography]==3.4.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
python-socketio[client]==5.13.0
    # via locust-cloud
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   mkdocs
    #   mkdocs-get-deps
    #   pymdown-extensions
    #   pyyaml-env-tag
    #   transformers
pyyaml-env-tag==1.1
    # via mkdocs
pyzmq==26.4.0
    # via locust
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   docker
    #   huggingface-hub
    #   kestra
    #   locust
    #   mkdocs-material
    #   python-socketio
    #   transformers
rsa==4.9.1
    # via python-jose
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements.in
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
torch==2.7.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.52.3
    # via sentence-transformers
typing-extensions==4.13.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   pyee
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via
    #   docker
    #   geventhttpclient
    #   requests
uvicorn==0.34.2
    # via -r requirements.in
watchdog==6.0.0
    # via mkdocs
websocket-client==1.8.0
    # via python-socketio
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   flask-login
    #   locust
wsproto==1.2.0
    # via simple-websocket
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
